[merb-core] Fix multipart POST for other encodings than ASCII-8BIT
Rack gives us POST in ASCII-8BIT only nomater what is encoding of
the request. This patch tries to brute-fix-it by force encoding on
the value of the parameters sent in the post.

File parameter is not affected because it's stripped from the
parameters before we start with text prameters.